Extract shared hover lock utility and simplify placement code
- Add hover-lock.svelte.ts with create_hover_lock() and create_dimension_tracker()
- Use stable object reference pattern for Svelte reactivity tracking
- Fix grid_resolution division-by-zero by clamping to minimum 2
- Export rects_overlap from layout.ts for test reuse
- Remove duplicated hover lock logic from ScatterPlot, BarPlot, Histogram
- Remove unused transform/position shim from placement results
- Merge placement-stability.test.ts into plot-legend.test.ts
- Remove stale orphaned comment in Histogram.svelte
